Prioritise class-like things as well as namespaces
Summary:
When searching by string, we can sometimes get a lot of results.
One way to improve relevance is to prioritise searching for containers (namespaces, classes).

We've been only doing this for "namespaces" in the past, but this was a bit restrictive. A simple, harmless improvement is to add classes/traits/ and other container-like things.

This ensures high-value symbols are likely to be in the results, even if there are 100s of constants with the same name.

Reviewed By: phlalx

Differential Revision: D40877322

fbshipit-source-id: 54f6dad19ed7f3318041d8e8abde790ccdb778a5