Retry server queue timeouts
Summary:
Server queue timeouts are short by design to protect the server from Thrift OOMs, and they are to be expected when sending fact batches concurrently from multiple hosts. In those cases, it's reasonable to retry immediately a couple of times.

To mitigate the chance of still overloading the server when all the concurrent clients retry, sleep a random backoff time.

Reviewed By: simonmar

Differential Revision: D40944796

fbshipit-source-id: 0aaed14d8bf904b891a9b3e4a0334188edc5a4db