Implement glass-side `method overrides` relation discovery
Summary:
This is a (mostly) quick to implement implicit method overrides discovery for Hack.

According to  https://docs.hhvm.com/hack/classes/inheritance#overriding-methods methods override by name. Glean only records those with explicit Overrides attributes, which is insufficient.

We actually want to:
- hide all inherited names that have an equivalent local name
- update those local names with a reference to their extended method-level parent

Do so just for Hack. The real solution will be to use the hack indexer to record a new implement MethodOverrides fact that we can then filter and search on. This is enough for now to get us moving along with correct API members however.

This will unconditionally remove exact local name matches from a container's inherited symbols if there's a local name in the container (e.g. consts, methods). That might be a bit too aggressive.

Reviewed By: simonmar

Differential Revision: D41297051

fbshipit-source-id: 19a8164607dcf969b926c169cf25a9ff23e3d17e