Inherited scope terms can be partially compiled to Angle
Summary:
Compile inherited scoped search queries into a suspended pair of the unevaluated scope term argument, and the rest of the angle code.
We want to treat something like "GK2ProdEvaluator::getCachedResult" as a query for:

- any parent of "GK2ProdEvaluator" (recursively, via inheritance)
- that contains exactly "getCachedResult"

to generate something like:

> codemarkup.search.SearchByScope { name = "getCachedResult" , scope = ["GK2ProdEvaluator"]|["IGK2Evaluator"]|["GK2EvaluatorWithGetNoABTestEvaluator"]|["GK2EvaluatorWithCache"] }

(note: even then we'll probably want to apply the inheritance "choice" rules in case there are N results, but tbd).

To do this we need to evaluate the rest of the search terms (case, language etc), parse out the scope, then compile to Angle.
Except for the scope term, which we'll compile _later_ as a disjoint set of scope terms, after looking up the root entity and querying its parents.

Note: We don't need cost tags , also, so removing those here. We were defaulting to 15s for "fast" queries and 5s for "slow" queries. We can just default to 5s for all search queries.

This diff just does some more plumbing and returns [] for the inherited search case for now. No change in behavior yet.

Reviewed By: simonmar

Differential Revision: D41395670

fbshipit-source-id: 337cb6b10565091224ade7be26f14326095ef466