Optimize disjoint parents query
Summary:
The disjunction of parent scope query blows up. We can do two things:

- uniq the set (i.e. remove multiply-occuring common answers.

For EntIntlLocaleSet this reduces the parent name terms from 673 to 208.

The bigger issue is the allocs from the nested disjuction inside the predicate. Float it out and name it.

This lets us search very large hierarchies quickly!

Reviewed By: eguitarz

Differential Revision: D41456510

fbshipit-source-id: 20312693f3e7b75b4da934617e57fa31fe66a066