Filter names in topological order outwards from base
Summary:
This implements the recursive name filtering, so we get exactly one symbol visible in the API. The first found outwards in topological order.

This is _close_ but not quite enough. We actually have to visit classes first, then traits, and finally interfaces. At the moment it is in breadth first parent order.

when updating 'Overrides' vs 'Implements' we need to know in the client if something is a parent trait/class or interface.

Reviewed By: simonmar

Differential Revision: D41574027

fbshipit-source-id: 6b591d374266e18c7ceed45c8e9ae24a5a4c427e