Plumbing for recursive name filtering
Summary:
To implement general name overrides/inheritance-based filtering of what methods are in scope we need to do our "local name shadows inherited name" recursively, in topological order, up the DAG.

So if we import a symbol via a class, and via a trait via a class, the inner class method will win.

To start, keep the topological ordering of the inheritance DAG around. Using a hashmap as the symbol id keys are likely to have long common prefixes.

Reviewed By: nhawkes

Differential Revision: D41570691

fbshipit-source-id: 68e8c7dba192634300524c47bb4c1b8340875d83