Stacked incremental DBs
Summary:
A stacked DB might induce changes to the ownership of facts in DBs
below it in the stack. So we have to propagate ownership from facts in
the stacked DB to facts in the base DB(s).  The basic approach is

* Each DB stores ownership information about its own facts and
  potentially facts in base DB(s).

* `getOwner()` in a stacked DB traverses the stack downwards,
  returning the first owner found.

* Therefore, if a fact F has owner A in the base DB and a new owner B
  in the stacked DB, we have to assign it owner `A || B`.

* Thus, an owner set in a stacked DB (`A || B`) may refer to owner
  sets in a base DB (`A`).

* A derived fact in the stacked DB will be given the correct owner,
  because its ownership is derived by calling `getOwner()` on each of
  the facts that it was derived from.

There's one gap in this plan, namely that when ownership is propagated
to facts in the base DB we don't have a way to propagate that to
derived facts too, because we don't keep track of the facts from which
a derived facts was derived, and to do that would be expensive. I'll
deal with this in later diffs.

Reviewed By: nhawkes

Differential Revision: D41520039

fbshipit-source-id: 89ed473e80b7853e95f8b6823e11c4701aa4f2e2