Keep the anchor entity when doing an inherited search for 'view context'
Summary:
The idea here is that when someone searches for TypeAssert::coerce and we succesfully find it inherited from TTypeAssertHelper::coerce we should not just dump the user on www/php/TTypeAssertHelper/coerce.

Instead, show them that symbol _in the context of class TypeAssert_. I.e. there's a symbol found, and a context container from which to view it. This does the server-side plumbing. Still need to update the bunny redirect and codehub search page to use this information.

Reviewed By: eguitarz

Differential Revision: D41676494

fbshipit-source-id: 72e3da2d15d24885261c03085e33014f8364e0b1