When doing an inherited scope search keep the anchor parent handy
Summary: We can use this to implement the `cp` behavior where Foo::bar resolves `Baz::bar` when inherited, but anchors the return to its location in `Foo`. So we'll return a secondary base symbol in these cases, so the client can note that the term was re-interpeted as a anchored inheritance search.

Reviewed By: simonmar

Differential Revision: D41643133

fbshipit-source-id: f71112b5621590de23f67a19c975a697e116329a