Fix duplicate syms issue in name resolver
Summary:
We weren't filtering names that shadow from sibling nodes. So if a class used a trait and another class, and both defined "f", we'd leave both in. rather than filtering class "f" by trait "f".

Was previously reported as T139543322

But actually debugged by specifying the behavior in a unit test

Reviewed By: simonmar

Differential Revision: D41866152

fbshipit-source-id: 28bbc178f22c02959a9c39a08a7be0701c6428ba