Handle ScopeOnly searches better
Summary:
Queries like `vec\fb\` or `HH\C::` are parsed as "scope only" queries. They are very likely to be containers, or namespaces.

Use this knowledge to search for them as a scope term. e.g. "fb" in "vec". Previously we didn't know what to do and just searched for "fb" which is less useful.

Reviewed By: eguitarz

Differential Revision: D41989665

fbshipit-source-id: d57e6668e5e5e72612969b19397a6047f6c9bcf3