indexing of callee targets
Summary:
Before this diff, indexer only indexes the span of a "callee" expression.

With this diff, it also indexes the xref target when it can be extracted from the callee expression. For instance xrefs to `foo` in expressions such as `foo`, `T::foo`, `XXX->foo`.

This simplifies client work which otherwise had to parse the callee expression and lookup in FileXRef.

Reviewed By: donsbot

Differential Revision: D42279043

fbshipit-source-id: 761e962892fcd5c9f6190c7979dbdf3c01837824