Optimise slice() for large exclusion sets
Summary:
We want to compute whether a set S is visible or not in roughly O(S),
giving us the total time to compute the visibility of all sets as
O(NS) where N is the number of sets.

`emptyIntersection()` was O(U) (where U is the number of excluded
units) because we were iterating linearly over U, which is bad when U
is large (I saw |U| = 65k in the example in the test plan).

Reviewed By: phlalx

Differential Revision: D42389483

fbshipit-source-id: 5233b232bb5f566f1371c03d0f92cb9acd9457f9