Make JS modules first class entities in Glean
Summary:
We've had a 'module_' case for JS in code.flow.angle for a long time but it was never implemented (as we didn't need it for source xrefs/find-refs).

It's useful however for:
- search
- api docs (container parent of something is a module)

We can treat modules as containers, search for them by name and associate a file location with them, all using the existing schema. We don't have spans for the file (though if we could sum src.FileLines we'd actually ahve the full span as well).

For now:
- make modules a symbol type
- add name search (based on flow.SearchByNameModule , and index of module names)
- add file + span { 0, 0 } locations shrug
- add basic symbol kinds (module, type, everything else)

To do the symbol kinds well we really need kind information in the decl/entity type in the Flow index. Have to know if the decls is a object, function, class etc. Can do it Hack-style, and have a ClassDeclaration (or ClassDefinition) predicate keyed by Declaration.

one interesting consequence is that file modules are now searchable (by filepath), but needs the full path. We might just get away with the file basename?

Module spans: we should index these. And presumably there's some sensible notion of span? or is it just 0, filesize?

Reviewed By: simonmar

Differential Revision: D42445956

fbshipit-source-id: ed62c672d49790e01d483bf490244cfc1fd05281