Resolve symlinks in fileFromEntry()
Summary:
Buck 2 is giving us relative paths to symlinks in buck-out for
tryGetRealPath(), which don't get resolved by anything currently.

The simplest solution I found was to replace `tryGetRealPath()` with
`std::filesystem::canonical(entry.getName())` and then use the
existing `goodPath()` to strip root from the resulting absolute
path. We have to leave the other existing logic in place, because
we don't *always* want to canonicalise the path since it might be a
symlink pointing outside of root and we don't want to resolve those.

Reviewed By: nhawkes

Differential Revision: D42476906

fbshipit-source-id: f39542b0ed7bf78760608a27f5e5129562b2e557