change fact owner cache page size to 4096
Summary: The idea here is to reduce the tail latency penalty for queries on an incremental DB stack. If a query misses the fact owner cache it has to load a page of the cache; this reduces the page size and hence the latency to load it. Reducing the page size will also help the overhead for the binary search we have to do on the contents of the page.

Reviewed By: phlalx

Differential Revision: D42640708

fbshipit-source-id: f223953dc57dce5356d4b551d0a3a3e6d3e99a57