index receiver static type
Summary:
When generating a Filecall fact for a method call, we also index the "receiver type" for the receiver of this method call.

This static type is determined as follows
1. find the `SymbolOccurrence.t` value for the method symbol. It was stored in the XRef map when processing xrefs in a previous stage.
2. get the `receiver_class` string from this `SymbolOccurence.t`  (see https://www.internalfb.com/code/fbsource/[3345feefcc47]/fbcode/hphp/hack/src/utils/symbolOccurrence.mli?lines=29)

I *think* receiver_class is what we need here, but need feedback from hh expert to make sure this is indeed what we need.

Pehraps Wilfred or madgen can help here (you can also look at https://fb.quip.com/uTawAjRPflxs or the linked task for more context)

Differential Revision: D42369572

fbshipit-source-id: 48d48e293ea66edeec9eb8262c0bd1d555560bc6