Fix interaction of manual and automatic derivations
Summary:
This allows us to mix automatic, version-less evolutions with source code evolution directives.

The types are now a bit more explicit and tell a clearer story:

- evolutions from version-less migrations: `Map PredicateRef PredicateId`
- evolutions from source code annotations: `Map SchemaRef SchemaRef`

We can turn that into a relationship between `PredicateId`s with a `Map SchemaRef [PredicateId]`.

We can take the database's content into account when deciding on the final evolution by having a function `(PredicateId -> Bool)`.

Reviewed By: simonmar

Differential Revision: D42607700

fbshipit-source-id: dace40822d589b7dc61bfa4f6df27c7a3643a98b