codemarkup.EntityDigest
Summary:
Adds a new predicate that maps entities to a digest produced by an unspecified hash function over the source code or a derivative of it. This can be then be used to track changes.

The primary use case is diff sketches with enough resolution to detect changes to function bodies and other constructs.

Since this is language agnostic, the predicate lives in `codemarkup` and annotates `Entity`es because that is what the diff sketcher uses  for top level declarations.

Reviewed By: simonmar

Differential Revision: D42776945

fbshipit-source-id: c6e4da03d783015915975cbe21e80094a34e0ed3