Add inner declarations to ClassDefinition and ModuleDefinition
Summary:
We do not explicitly store information about class/instance methods and variables.
We need to store it to show this information on Docs API pages.
In this change we're adding.
- methods
- instance variables (class attributes)
- class variables (static class attributes)

Tricky moment:
It's designed that use the same predicates to store information about all the functions and variables: class scoped, local, global, etc.
So (in my understanding), we don't have a special place to store information if a class variable is class scoped or instance scoped.
In this RFC I chose to store this information using different "bucket" in class definition.
However, we can introduce special an extra field in VariableDefinition, or create a Method/Attribute predicates (like we do in Hack) or do it in another way.

UPD: Added
inner_classes to ClassDefinition
declaration to ModuleDefinition

Reviewed By: simonmar

Differential Revision: D42861049

fbshipit-source-id: 8c017ee1333f416975fc0010628d6e51ef8e123c