Don't use Thrift serialization (#314)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/314

I want to eliminate the dependency on the C++ Thrift compiler, which will make open-source Glean much easier to work with. This diff eliminates our usage of Thrift serialization in glean/rts, which is the biggest roadblock to getting rid of the dependency.

Serialization is mostly done using the existing `binary::Input` and `binary::Output` types, with a little library on top (`serialize.h`) to provide standard serialization for a few types and containers like `std::vector` and `std::optional`. I didn't make it super fancy because it didn't seem worthwhile, but definitely open to suggestions for improvements.

Reviewed By: nhawkes

Differential Revision: D42844201

fbshipit-source-id: 889b147b3ca31e62e677a95e2c8a8ed9dfe3d7cb