Get rid of all Thrift C++ codegen
We don't depend on any gen-cpp2 source files in the OSS build any
more, so we don't need to run the thrift1 compiler. Lots of cleanup
and builds should be faster.

The remaining dependency on fbthrift is the client/server
interaction. We could switch that to use
`Thrift.Channel.SocketChannel` instead.