Detect co-recursive derivation queries
Summary:
We don't support recursion at the moment, so we want to detect co-recursive definitions
at schema resolution time.

For default derived predicates, this is tricky.

Whether they will in fact recurse or not depends on the db's content and we already have co-recursive definitions to allow for two versions of the same predicate to be available (in `buck.File`).
With our schema migration options I think we probably don't need to support co-recursive default derived predicates any longer. This diff does allow them, but I would be in favour of removing support for this.

Reviewed By: simonmar

Differential Revision: D42844997

fbshipit-source-id: 5882c60942771ceb294b25267b4fc2ed6fff60e9