binary to precompute queries
Summary:
Motivation: Phabricator Diff navigation. See https://fb.quip.com/Vu8GAJ1n69xY for the big picture.

This diffs adds a new binary `glass-snapshot`.

When invoked with a list of files and an output file, It generates the `DocumentSymbolListXResult ` for a the list of files, serializes it and stores it to a file.

Reviewed By: simonmar

Differential Revision: D43057560

fbshipit-source-id: 767b8f4dc0792b74cfe8eb79cf936dd7dd938fc1