ODS counter to track leaked DBs
Summary:
The new counters are:

* glean.db.slack
* glean.db.<repo>.slack

These counters measure the amount of total and per-db "slack" in bytes. They will help us to understand how much space is wasted and could be reclaimed by fixing the DB leak.

The "slack" is the difference between the remote and local+remote retention sets. That is, the DBs that a server keeps locally despite not being in the retention set.

Slack is only justified when all other shards are still restoring newer DB instances, in which case the older DB instances are needed to prevent an outage. This happens very rarely with randomized sharding, and constitutes a resource leak that we track.

More details in this [Workplace post](https://fb.workplace.com/groups/csi.wtf/posts/5534304019979552)

Reviewed By: simonmar

Differential Revision: D43127933

fbshipit-source-id: 40acfc0b563c9d634af9b274ac31bca3e18212b2