More performance optimizations for searchRelatedNeighborhood
Summary:
This is a quick optimization of the SymbolDescription reponse for the inherited children set of searchRelatedNeighborhood. We show this as a list of things that extend or inherit the class/container. We only ever show the name, parent, kind, type sig and symbol id.

e.g. we have been computing comment spans/docblocks for potentially hundreds of unseen symbols. Ouch.

We have been computing the full details, which is expensive on some classes.  We make this report scale a bit better by not computing some unnecessary things for these inherited chidlren symbols, without affecting codehub. It just reduces the computation and amount of data transferred to Hack.

Reviewed By: simonmar

Differential Revision: D43227376

fbshipit-source-id: e8b15d2491a509d2204167db04f5128ace394928