Get rid of all Thrift C++ codegen (#320)
Summary:
We don't depend on any gen-cpp2 source files in the OSS build any more, so we don't need to run the thrift1 compiler. Lots of cleanup and builds should be faster.

The remaining dependency on fbthrift is the client/server interaction. We could switch that to use
`Thrift.Channel.SocketChannel` instead.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/320

Reviewed By: lazamar

Differential Revision: D43186042

Pulled By: simonmar

fbshipit-source-id: b1e511d24b7065084ff123bf7c862e5209761783