Flush the WAL before close
Summary:
We've seen that it can take a long time (1min+) to re-open an
incomplete DB if it was closed. This shouldn't normally happen, but
the clang indexer sometimes stalls for a long time during
derive-function-calls and the DB gets closed as idle, and then
re-opening it is slow because the WAL is replayed. Let's flush the WAL
when we close the DB to avoid that problem.

Reviewed By: lazamar

Differential Revision: D43271773

fbshipit-source-id: 072536e4a0b935cbe8f59b025ba68f840cbb15cc