More brief describe() optimisations
Summary:
We don't need two copies of the location (actually we don't use locations at all but seems useful to keep for debugging).

Later in the stack I'm going to break this out as a separate type for brief symbol descriptions so we can avoid parsing/processing lots of default values in structs.

Reviewed By: phlalx

Differential Revision: D43261306

fbshipit-source-id: 34e4c213caf1aeb08143ac0dfa59f4b6d9eecc1a