Generate type signatures for variables
Summary:
This avoids a problem where we can generate an Angle query where the
Angle compiler infers a different type from the one the client
expects.

If the context has a predicate type, the Angle compiler accepts an
expression of either the predicate type or the key type, and if it's a
variable then we infer the variable to have the predicate type.

e.g.

```
X where cxx1.FunctionName X
```

will be accepted by the compiler with `X : cxx1.Name` inferred.
However this DSL code

```
var $ \(x :: Text) -> x `where` [ wild .= predicate Cxx.FunctionName x ]
```

typechecks and generates the same query, with the result that the data
returned by the server doesn't match what the client is
expecting. Generating type signatures fixes the mismatch.

Reviewed By: lazamar

Differential Revision: D43195475

fbshipit-source-id: 8f61249b39ec09fbfc9107fbb7df187592376d6c