Retire java.{2,3,4,5} schemas
Summary:
All the indexes are on java.6 for a very long time now, which contains necessary copies of all relevant types. GC the old versions as they make working with the schema error prone.

The new continuous versioning system will retain copies of the old things until all the traffic dries up, so we can do this quite safely now.

Reviewed By: pepeiborra

Differential Revision: D43447851

fbshipit-source-id: a288f544a8cc572be87643c3c02693a836cbbd4f