Add missing xref clause to derivation of HackFileEntityXRefSpans
Summary:
It's technically feasible to generate these declaration references,
and we should be getting them for constant refs in the code - but
they're being ignored.  Don't ignore them.  Well, OK, ignore
Occurrences as those are method references that can't be resolved to a
caller type and we'd still like to get argument spans out if that happens.

Reviewed By: simonmar

Differential Revision: D43374511

fbshipit-source-id: 8ce61ead5386d80dc8c3ef9eea631c96c741592c