Implement UsrToReferences glean thrift api.
Summary: to support the Refs grpc call, we must be able to take any usr and get all the usages of the corresponding symbol. I initiallity tried using the exiting `clangUSRToDefinition`  (then calling the regular f`indReferenceRanges` on the returned symbol) however `clangUSRToDefinition` filters out when the usr is of the same location as the definition. So here we provide a new thrift rpc that is an algamation of the FindReferenceRanges and clangUSRToDefinition.

Reviewed By: simonmar

Differential Revision: D43334669

fbshipit-source-id: 5cc7efad774d3d58df773fb2359af61804210498