Expose docstring content from index along with comment spans
Summary:
This generalizes the codemarkup.EntityDocumentation call to include optional docstring content where it exists. It should be stored in the index as markdown text, that can be passed through to CodeHub's markdown renderer without alteration.

We continue to return the SCM file/span for compatibility in all cases.

For perf, I'm intentionally combining the two comment span predicates (span/file and text) so we don't have to make a new call from glass. Also check the predicate gets the order of the python.DeclarationLocation lookup correct (want to run _after_ we know there's a text span for the decl).

Reviewed By: nhawkes

Differential Revision: D43449433

fbshipit-source-id: 3a0426d8732ea783af0cb5c11009082dea2a9ff6