Support entity documentation stored natively in the indexer
Summary:
Extend the glass comments attribute to include the span of the origin of the comment, and
the rendered content (in markdown or doxygen format, currently). Just makes some slots and types.

We will also need:
- xref spans relative to the pretty content, and the symbol ids
- an array of doxygen-like directives if they exist (e.g. param,use)

the content should be a valid Facebook markdown string (as defined by hack and js libraries that parse markdown)

Reviewed By: nhawkes

Differential Revision: D43448731

fbshipit-source-id: bb635af53c2cb8b0d5edbf911e17a060fa84a348