Use stable unique
Summary:
At the moment we randomise repo order, which randomises repos hash and causes bad things

Data.List.nub is quadractic, but repos is small, should be fine

We should really pick the latest repohash rather than the first, but that's an exercise for later. We can always just reorder db names likely to be fresher to be first in our repos list config.

Reviewed By: donsbot

Differential Revision: D43470300

fbshipit-source-id: 71c2ddd4c3129d62b2485f602f911b51f1e5b99b