handle duplicate entity locations
Summary:
If the database contains multiple locations for an entity, the deriver will try to define the same `codemarkup.EntityDigest` fact twice and the Glean write server will throw an exception. So when can this happen?

* In the Python case, this can happen any time a source file redefines a declaration. The linter will probably catch this, but there might be corner cases.

* In Hack, this seems to happen even without re-declarations for reasons unknown in a very tiny fraction of diffs. See task comments for more details.

The workaround is to define an `EntityDigest` fact for only one of the locations. We pick the earlier starting one to minimize the chances of picking an invalid location.

Reviewed By: simonmar

Differential Revision: D43446264

fbshipit-source-id: 649a0d43c0523cf5f0c2e2e5f8475684e19061a1