group all writes
Summary:
Handle duplicate locations for a single entity by using the earliest one in order to avoid a "invalid fact redefinition" error.

For Hack, these duplicate locations are being caused by having two `hack.DeclarationSpan` facts for the same declaration. It's unclear what causes this, but we do not want to crash the diff sketcher - instead emit a warning.

Group all the writes for better performance: 30s -> 12s for the example at T145559263

Reviewed By: phlalx

Differential Revision: D43446267

fbshipit-source-id: bc530d3a5785b907e966082952259ff5f3f3cdde