benchmarking regression test
Summary:
Benchmarking target to conveniently measure the cost of running the digest standalone deriver.

The implementation reuses the glean regression test framework to setup a Hack DB containing ~10000 declarations and then run the deriver on it, printing time and allocations to stdout.

I opted for Hack instead of Python because the Python indexer was much slower to index a file with 2000 declarations.

Reviewed By: phlalx

Differential Revision: D43446270

fbshipit-source-id: 854a212d8872f76ccc94ab82c3586a7aa99cd993