Avoid seeks based on a fact's references
Summary:
This is the smartest trick we apply and it may be useful to speed-up our writing flow too.
The main idea is that if we know a fact cannot be in the db/substitution we do not search for it.

If during parallel deduplication a fact references an Id we have no substitution for, then there is no point in seeking for it in the db. So we just skip this one and queue it for the final deduplication.

If a fact references the id of a fact that we know for a fact is not in our database, then we can already make the call that this fact is not in our database.

In this diff we optimise these cases by differentiating a missing substition from a substition to an invalid id.

With this change we go from 12% to 92% of facts being dealt with in the parallel deduplications.
This means that the single-threaded deduplication is only handling 8% of the total facts.

With that we are now bottlenecked on the reading speed, which is addressed in the next diff.

Reviewed By: simonmar

Differential Revision: D41549769

fbshipit-source-id: 719da1dfeb1425b7611eb76ac642a137b5ae0861