Faster and safe slices
Summary:
This is an algorithmical improvement that doesn't rely on unsafe primitives (re D43446265). Instead of calling `slice` for every range, sort the ranges and drive a pointer traversal over the file contents.

Performance is good for the example, almost as good as the unsafe slice version. I expect it will be even better once we upgrade to text 2.0

Reviewed By: simonmar

Differential Revision: D43574148

fbshipit-source-id: 8231b81b2f0f7e2c520807b4603d2bbc1e45c77a