Better reordering around disjunctions
Summary:
If a disjunction fails to resolve, we currently don't try reorder
statements around it. Doing this is difficult in general, because it's
not possible to predict whether a disjunction will resolve or not
without trying. So that's exactly what we do: try to resolve the
statement, if it doesn't work then put it to the back of the queue and
try the next one. More details in the comments.

I think this should elimiante a chunk of the spurious "unbound
variable" errors that pop up with complex derived predicates.

Reviewed By: lazamar

Differential Revision: D43502359

fbshipit-source-id: eeb3a3e971f76fd35785911d3674cf6e9fdd0b6d