Restrict structural comparison to tuples
Summary:
Structural comparison of records can hide type errors. Here is an [example](https://www.internalfb.com/diff/D43376125?transaction_fbid=741598830672241).

For backward compatibility we bump the angle version.
We make `eqType` consider the angle version and reserve the field names we use for tuples so that they are not used in user records.

Reviewed By: simonmar

Differential Revision: D43437585

fbshipit-source-id: 00f92c474f0df5b3bc38e77ea09c0ae7cb837855