Index query transformations by schemaId
Summary:
We want the evolutions applied to a query to be the ones coming
from the union of the db's stored `ProcessedSchema` and the client's `ProcessedSchema`.

To do that we will use the `SchemaId` used by the query to determine the
`ProcessedSchema` that should be unioned with the stored schema.

Reviewed By: simonmar

Differential Revision: D43535557

fbshipit-source-id: e3282d829078cd4bf64c3f9642e2848932c81fc8