Refine the attribution of pasted macro ranges.
Summary:
This is a follow-up to D32015488. In that diff, we made it such that given an example like this:
```
#define PASTE(x,y) x##y
#define MACRO(x) x

void foobar() {}

void f() {
  MACRO(PASTE(foo,bar))();
}
```
The xref to `foobar` would be coming from `PASTE`, rather than `PASTE(foo,bar)`. This diff makes it such that the xref to `foobar` comes from `PASTE(foo,bar)`.

This is more accurate than the previous state in an abstract sense, but it's also important from a practical view of wanting `PASTE` to jump to the definition of `PASTE`.

Reviewed By: simonmar

Differential Revision: D43736111

fbshipit-source-id: 3b253386728fb288642883f3e7e6d2e2db02b48a