Replace uses of `rangeOfToken` and remove it.
Summary:
This diff removes the use of `rangeOfToken` function and removes it.

`rangeOfToken` is a function that returns a `SourceRange` that spans a token given a `SourceLocation`. However, the way Clang handles `SourceRange`s is such that a single location implicitly refers a token at that location. Rather than explicitly computing the `rangeOfToken`, a `SourceRange(location)` will be resolved to become the token range later on via `SourceManager::getExpansionRange`.

Moreover, the `rangeOfToken` function currently skips doing anything if the range is a macro ID. We shouldn't skip doing this work for macros, as we still do want the token range within the macro. Having a `rangeOfToken` function separately gets in the way more than it helps us.

Reviewed By: simonmar

Differential Revision: D43736113

fbshipit-source-id: 6edc64ddd09bec97e91c55a83eb77bb9de7276a3