Rewrite of reordering logic
Summary:
The previous reordering logic had bad behaviour in some cases (see
`lang/clang/tests/regression/perf/decl/functionDeclaration.query`) and
it was also quite difficult to improve it.

In this diff I'm replacing it with a greedy algorithm that can make
better choices in some situations. Importantly it does the obvious
stuff unconditionally: choosing O(1) statements whenever it can, and
only makes an approximate judgement when it has to.

Reviewed By: lazamar

Differential Revision: D43983209

fbshipit-source-id: 2c3c7e439756cd6a1a0a9c4659849458735acace