Consolidate buck schema
Summary:
Rather than having predicates scattered across `buck.1`, `buck.2` and `buck.3`,
we now move everything to `buck.4`, which evolves `buck.3` but does not inherit
from it.

There are no changes to the predicates.

This is the first instance of a [schema squeeze](https://fb.workplace.com/groups/glean.python/posts/5617095165026775). The recent fixes to the interaction of evolution annotations and version-less migrations should make this work.

Reviewed By: nhawkes

Differential Revision: D43833892

fbshipit-source-id: 548be2e8d3c2635cae6abbe7e1f3496d2abf6c0a