Tweaks to the reordering logic
Summary:
Make array generators have the same cost as a prefix match. It's just
guesswork really, but that seems kind of reasonable: an array is
probably smaller than a whole predicate.

This makes this problematic query work

```
X0 where X0=codemarkup.EntityReferences.30 {target={cxx={decl={function_=($415491411:cxx1.FunctionDeclaration.5)}}}, file="fbcode"..}
```

Before reordering we have

```
{X0_0,{(0|(0|(5|{1051}))),File_3,(0|Span_6)},{}} where
  [File_3 = src.File.1 "fbcode".._56;
   (
     [_57 = cxx1.TargetUses.5 {(0|(5|{1051})),File_3,Uses_12}];
     [Spans_13 = prim.relToAbsByteSpans Uses_12];
     [Span_6 = Spans_13[..]]
     );
   X0_0 =
     codemarkup.EntityReferences.30<- {(0|(0|(5|{1051}))),File_3,(0|Span_6)}]
```

The `File_3` statement is a prefix match, and the group below it contains `Spans_13[..]` which iterates over the array. If we consider that to be a prefix match then we choose the correct ordering, because `File_3` is bound by the group so we put the group first.

```
{X0_0,{(0|(0|(5|{1051}))),File_1,(0|Span_2)},{}} where
  _ =
    cxx1.TargetUses.5
      {(0|(5|{1051})),File_1:src.File.1,Uses_3:[src.RelByteSpan.1]};
  Spans_4:[{ tuplefield0 : nat, tuplefield1 : nat }] =
    prim.relToAbsByteSpans Uses_3;
  Span_2:{ start : nat, length : nat } = Spans_4[..];
  X0_0:codemarkup.EntityReferences.30 =
    codemarkup.EntityReferences.30<- {(0|(0|(5|{1051}))),File_1,(0|Span_2)};
  File_1 = src.File.1 "fbcode".._
```

**However** this is pretty fragile so I don't think we should rely on it. When I tried to add a perf test for this there were cases that didn't work, when `codemarkup.EntityReferences` expanded to something more complex (because it depends on which facts are in the DB).

Reviewed By: donsbot

Differential Revision: D44140256

fbshipit-source-id: 5280225a85f1e6fcc08195ae0628f1e292a07862