Allow missing evolutions for types without facts
Summary:
If schema `x.2` evolves `x.1` but there are no facts of either schemas then we
won't compute the transformation for any of their predicates.

But that doesn't mean that predicates from some other schema `y.2` which
mention `x.2` can't evolve predicates from `y.1` which mention `x.1`.

The test case demonstrates the issue. Currently that case is a validation
failure, which is what caused S330014. This diff addresses that.

Is't it a problem that query patterns won't be transformed? Could this cause
unexpected decoding issues?

I don't think so. If there is an evolution the code generated for a
`FactGenerator` will use an evolved pattern to search with the evolved
predicate type and transform the facts found back into the type used in
the query. But if there will be no results we might as well search using the
original pattern and the original predicate type and get the same 0 results
which we won't need to transform back.

Reviewed By: simonmar

Differential Revision: D44205376

fbshipit-source-id: 045b663c5fe6797e4a8bad8ca4576bbf6a810df2