Implement case-insensitive scope/namespace search for C++
Summary:
We've not been doing case-insensitive scope search in C++ which is a shame and a bit dumb.

E.g.
> folly::random::rand32
won't match
> folly::Random::rand32

Use the same trick as Python to build a scope term converter keyed by the insensitivity flag.
As we compile the query we check the value of the flag and either use the lower case table or not to find the cxx1.Name in each scope field. Not too bad.

Staged to avoid any issues:
- adds 3 new 'case aware' search predicates
- deletes the old ones in the next diff

Example of what we need to hit:

Reviewed By: nhawkes

Differential Revision: D44333045

fbshipit-source-id: 66c639e352592f0acdeeb2869dc1e572f333327c