Stage removal of briefer description Map.union
Summary:
As specified in previous diff, we have some parts of the details report that are now 'brief' and moved to a different field in the thrift response. We can drop them here from the main map.

We can push this once the thrift is in www and the client side is switched out

Reviewed By: pepeiborra

Differential Revision: D44326343

fbshipit-source-id: 602e1da9ec5a824652c75855d786d56f0488a835