Propagate bit indicating results were truncated somewhere
Summary:
For document symbols we do want to keep track of whether Glean or Glass had to truncate the result.
Instead of throwing the individual flags away, retain them and hand the client a True if any of the underlying queries for defs or xrefs had to truncate (or Glass itself if the sum was > $limit).

We can also log this once the logger config is updated from D44410231

Reviewed By: simonmar

Differential Revision: D44410443

fbshipit-source-id: e750f9a91f371e8e19d273e93f757e048662e777