SendAndRebaseQueue: don't deadlock when the sender dies
Summary:
We're getting deadlocks from SendAndRebaseQueue in some cases when
there's an error from the write server.

The SendQueue does propagate exceptions properly when one of its
threads dies, so that's not the problem. But here we're deadlocking in
`deleteSenderPool` during cleanup, because there was an exception
that prevented the Sender from being put back on the `srqSenders`
queue.

It will be non-deterministic whether we see the
BlockedIndefinitely exception or the original exception, because it's
up to `Async.concurrently` which one gets propagated, so I couldn't
make it happen reliably in a test. Still, this should fix it.

Reviewed By: lazamar

Differential Revision: D44744800

fbshipit-source-id: 896e4d2aab97a148e7ee340a70c3d2b710a12b9c